refactor: removes the use of the uuid package
